move logger to __init__.py and rename to mantis
